Restructured the commit code
