+++ app/controllers/thingy-controller.js
@@ -1,43 +1,4 @@
 const Thingy = require('../models/thingy-model');
-const Mqtt = require('../mqtt');
-
-/**
- * Listens to the connection status of the Thingies. When true, the Thingy just connected,
- * when false, it disconnected.
- */
-Mqtt.client.subscribe('+/Connected');
-Mqtt.client.on('message', async function (topic, message) {
-    if(topic.endsWith('/Connected')) {
-        // get MAC of thingy that threw the message
-        var thingyMAC = topic.replace('/Connected', '');
-        // get Thingy if available
-        var thingy = await Thingy.findOne({macAddress: thingyMAC});
-        // if no exists, add to DB
-        if(thingy == null) {
-            thingy = new Thingy();
-            thingy.macAddress = thingyMAC;
-        }
-        // process a connected Thingy
-        if (message == 'true') {
-            // Thingy is now available
-            thingy.available = true;
-            try {
-                await thingy.save();
-            } catch (err) {
-                console.error(err);
-            }
-        // process a disconnected Thingy
-        } else {
-            try {
-                // Thingy is no more available
-                thingy.available = false;
-                await thingy.save();
-            }catch (err) {
-                console.error(err);
-            }
-        }
-    }
-});
 
 /**
  * Find all thingys.
@@ -52,8 +13,11 @@ async function findAll(ctx, next) {
     // thingy.macAddress = '11-22-33-44-55-66';
     // thingy.lockedForUser = null;
     // await thingy.save();
-
-    ctx.body = await Thingy.find({});
+    let available = ctx.request.query['available'];
+    if(available !== undefined && available === '1')
+        ctx.body = await Thingy.find({available: true});
+    else
+        ctx.body = await Thingy.find({});
 }
 
 /**
@@ -109,7 +73,6 @@ async function unlock(ctx, next) {
 
 module.exports = {
     findAll,
-    findAllAvailable,
     lock,
     unlock,
 };
