+++ app/routes/thingy-routes.js
@@ -4,7 +4,6 @@ const router = new Router();
 const Ctrl = require('../controllers/thingy-controller');
 
 router.get('/', Ctrl.findAll)
-    .get('/available', Ctrl.findAllAvailable)
     .post('/:thingyId/lock', Ctrl.lock)
     .delete('/:thingyId/lock', Ctrl.unlock);
