+++ open-api-specification.yaml
@@ -132,6 +132,12 @@ paths:
     get:
       summary: Get a list of thingys
       tags: ['Thingys']
+      parameters:
+        - in: query
+          name: available
+          schema:
+            type: integer
+          description: 1 if only the available thingys should be returned
       responses:
         200:
           description: Success
